fix(cuda): the Q5_K GEMV read each value's fifth bit from a sequential bitmask — every Q5_K tensor decoded wrong on the GPU (#3111)
In ggml's block_q5_K, the fifth bit of value l in sub-block s is bit s
of qh[l], i.e. byte val_idx % 32, bit val_idx / 32. Q5KGemvKernel read
byte val_idx / 8, bit val_idx % 8 instead. Its qs offset and nibble
selection were already right, so only the high bit was scrambled. That
was enough: on the FALSIFY-QDOT-009 golden block, the GPU dot against a
ramp came out 2.659 where gguf-py gives 0.622. Measured on an RTX 4090
by running the new test against the shipped indexing.

The only earlier GPU Q5_K test asserted `q5k_gemv(...).is_ok()` on
synthetic weights. FALSIFY-QDOT-010 runs the kernel on the golden block
from #3110, now a fixture shared via include!(). A one-hot activation
must return each of the 256 gguf-py values, and a ramp must reproduce
gguf-py's dot. Both tests pass with the fix and fail with the shipped
indexing.

Q5KKernel (the GEMM) has its own wrong layout: the qh and qs offsets
and a 12-bit scale unpack. Nothing in inference dispatches it; only
q5k_matvec in the layer-trace path and tests do. It stays tracked on
#3111.

Refs #3111

Pmat-Ticket: PMAT-1104
