+++ contracts/aprender/quantized-dot-product-v1.yaml
@@ -264,6 +264,15 @@ falsification_tests:
     with gguf-py on 5,959,751 of 6,291,456 values of blk.0.attn_qkv and its layer-0 QKV sum was
     +26.64 against llama.cpp's -17.06. The dequantizer and the fused dot shared the invented
     layout, so every test that compared one with the other passed
+- id: FALSIFY-QDOT-010
+  rule: The CUDA Q5_K GEMV decodes ggml's block_q5_K, judged against the same gguf-py values as
+    FALSIFY-QDOT-009, never against another aprender reader
+  prediction: q5k_gemv on the golden super-block returns each of its 256 gguf-py values for a
+    one-hot activation, and gguf-py's dot (1e-4 relative) for a ramp
+  test: cargo test -p aprender-serve --features cuda --lib -- tests_q5k_ggml (needs an NVIDIA GPU)
+  if_fails: The kernel takes a value's fifth bit from the wrong qh byte or bit. Found 2026-09-11
+    (#3111, PMAT-1104) - the shipped kernel read byte val_idx/8, bit val_idx%8 instead of byte
+    val_idx%32, bit val_idx/32; on the golden block its ramp dot was 2.659 against gguf-py's 0.622
 qa_gate:
   id: F-QDOT-001
   name: Quantized Dot Product Contract
@@ -273,7 +282,7 @@ qa_gate:
   - Format registry matches trait implementations
   - Dispatch table covers all formats
   - Cross-format isolation holds
-  pass_criteria: All 7 FALSIFY tests pass
+  pass_criteria: All 8 FALSIFY tests pass
   falsification: Introduce a sign error in nibble extraction — gate must catch it
 equations:
   identity:
