+++ crates/aprender-serve/src/cuda/executor/mod.rs
@@ -294,6 +294,10 @@ mod test_fixtures;
 #[cfg(test)]
 mod poison_trace_test;
 
+// FALSIFY-QDOT-010 (#3111): the Q5_K GEMV against gguf-py's values of a llama.cpp block
+#[cfg(test)]
+mod tests_q5k_ggml;
+
 // COV-003 through COV-006 (layer preload, kv_cache, attention, quantized)
 #[cfg(test)]
 #[path = "tests_cov003_preload.rs"]
